Be more consistent on package name.
We now rather rename the package at import time. While it’s possible to
have package name which differs from import path, this is an
antipattern. It’s better to be explicit.
Also, use the fqdn part for the import in tests as go modules prefers
this over relative ones.
Use this opportunity to import at the top the package we are testing against.