Fix deadlock when cancelling a command's context
When the context is cancelled, the error from the IO goroutines
was ignored anyways, so we can simply not wait for them.

They'll be killed by `c.closeDescriptors` a few lines below, so there
is no risk of leaking them.

The reason we cannot simply move `c.closeDescriptors` above the waiting
is that when cancel is not called the goroutiones may still be writing.
Closing them early may cause a panic.