Fix almost empty error when wsl.exe fails.
If wsl.exe doesn't exit 0, then we read the byte buffer to find "Wsl/Service/WSL_E_DISTRO_NOT_FOUND".

By doing so we're exhausting the reader, so the last line of this function essentially assembles an empty message, because
there is nothing more to retrieve from the reader.

The fix is write the buffer contents into strings and use those when comparing and generating error messages.