Fix TestShell timeouts
It's been a while that GoWSL CI has been shown issues with TestShell.

After some investigation I figured that first boot of the images used for testing is taking longer than what the test cases expect.
Thus the timeout.
From GoWSL's point-of-view it'd be better to use lighter images, without systemd for example.
But since GoWSL's main use case is with Ubuntu, let's keep the images as is and just give the tests more time to run.